More Refactors
Switching to predictive parsing. I will be attempting to add tests during a parse failure to attempt to identify specific syntax errors during parse time. This, due to my utilization of peeking, has led me to use a token stream structure that relies heavily on cloning. I may attempt to figure out something to reduce the amount of clones, but that is not a priority, yet anyways.